+++ changelog.md
@@ -1,5 +1,9 @@
 # Upcoming
+* Tiredness
+  * Phantoms can now spawn even at Tired II, but much less often
 * Rounded sprint speed on armor
+* Moved Copper items out of Minecraft namespace
+  * This should fix incompatiblities with other mods that add Copper items
 * Fixed Glow Block missing recipe and added an advancement
 
 # 5.10.0.1